update
this updated version of the JavaScript code, I added a response-text class to the response <p> element and updated the CSS to apply styling to that class. This should make the output text more visible and centered on the page. I also updated the references to the responseText property in the fetchChatGPT function to use the correct property name (responseText instead of responseText).